// projects/app-api/src/lib/api/models/baseOutputDeviceGroup.ts
/**
 * Dsd API
 * DSD Information
 *
 * The version of the OpenAPI document: 0.0.1-SNAPSHOT
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { ResponseStatus } from './responseStatus';
import { DeviceGroup } from './deviceGroup';

export interface BaseOutputDeviceGroup {
  errors?: Array<string>;
  message?: string;
  currentPage?: number;
  pageSize?: number;
  totalPages?: number;
  status?: ResponseStatus;
  data?: DeviceGroup;
}
export namespace BaseOutputDeviceGroup {}